Handle reversedKey inside StateTreeViewItem
